# docs/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CreateRepositoryRequest" : {
        "properties" : {
          "codeowners" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "defaultBranch" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "orgRepo" : {
            "type" : "string"
          },
          "serviceId" : {
            "type" : "string"
          },
          "topics" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "codeowners", "defaultBranch", "orgRepo", "topics" ],
        "type" : "object"
      },
...
      "RepositoryResponse" : {
        "properties" : {
          "codeowners" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "defaultBranch" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "language" : {
            "type" : "string"
          },
          "orgRepo" : {
            "type" : "string"
          },
          "serviceId" : {
            "type" : "string"
          },
          "topics" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          }
        },
        "required" : [ "codeowners", "defaultBranch", "id", "orgRepo", "topics" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/repositories" : {
      "get" : {
        "operationId" : "listRepositories",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/RepositoryResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "List all registered repositories",
        "tags" : [ "Repositories" ]
      },
      "post" : {
        "deprecated" : true,
        "operationId" : "registerRepository",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateRepositoryRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepositoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Register a repository in the graph",
        "tags" : [ "Repositories" ]
      }
    },
    "/api/v1/repositories/{id}" : {
      "delete" : {
        "operationId" : "deleteRepository",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "summary" : "Delete a repository from the graph",
        "tags" : [ "Repositories" ]
      },
      "get" : {
        "operationId" : "getRepository",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RepositoryResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "summary" : "Get a repository by ID",
        "tags" : [ "Repositories" ]
      }
    },
...
  },
...
}